iss #24: extract config readers to package
