+++ core/src/main/scala/ru/maizy/ambient7/core/config/reader/UniversalConfigReader.scala
@@ -1,4 +1,4 @@
-package ru.maizy.ambient7.core.config
+package ru.maizy.ambient7.core.config.reader
 
 /**
  * Copyright (c) Nikita Kovaliov, maizy.ru, 2016-2017
@@ -13,6 +13,7 @@ import com.typesafe.config.{ Config, ConfigFactory }
 import configs.Configs
 import scopt.OptionParser
 import ru.maizy.ambient7.core.config.helper.ConfigRuleOps.IfSuccessOp
+import ru.maizy.ambient7.core.config.{ Ambient7Options, ParsingError, configLogger }
 
 object UniversalConfigReader {
   type CheckResult = Either[ParsingError, Unit]
